Added getCreditTransactionByInterval method to ctRepo, covered it with tests

+++ src/main/java/varta/dto/EnrichedTransactionDto.java
@@ -4,12 +4,14 @@ import java.math.BigDecimal;
 
 public class EnrichedTransactionDto {
     //velocity features
-    private Double velocity1H;
-    private Double velocity24H;
+
+    private int velocity1H;
+    private int velocity24H;
     private Integer distinctMerchants1H;
     private Double failRatio1H;
 
     //monetary deviation features
+
     private Double avgSpend30D;
     private Double zScore;
     private Double ratioToMedian;
